Explain the purpose of this project
It's always important to explain what we are doing and how it works.
By writing the README file before any scripts, we focus on user
experiment more than on the development.

The goal is simple: user download a bootstrap script and they just
have to follow instructions (fewer as possible) to have their
operating system ready.