First version
The boostrap script detects your operating system and switch to the
corresponding branch if it exists.

For the sake of simplicity and because I only want to support ubuntu
for now, this script relies on lsb_release (some other operating
systems use lsb-release other do not have it pre-installed).
It also relies on unzip to extract archive.

If we include other operating systems, this script will most likely
ensure to use appropriate binary according to detected operating
system.